fixes for time_sensor_broadcaster ready for tests

+++ time_sensor_broadcaster/include/time_sensor_broadcaster/time_sensor.hpp
@@ -62,7 +62,7 @@ public:
     get_seconds();
 
     // update the message values
-    message.data = seconds_;
+    message.data = static_cast<float>(seconds_);
 
     return true;
   }
